rename interactor to FocusableInteractor
